pyproject.toml
[project]
name = "cer-django"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "cookiecutter>=2.6.0",
    "django>=5.1.7",
    "django-dump-die>=0.1.7",
    "django-environ>=0.12.0",
    "django-extensions>=3.2.3",
    "django-tailwind[reload]>=3.8.0",
    "django-unfold>=0.52.0",
    "gunicorn>=23.0.0",
    "pillow>=11.1.0",
    "psycopg[binary]>=3.2.6",
    "requests>=2.32.3",
    "whitenoise>=6.9.0",
]
